refactor: some changes in ft_strcmp
Rearranged how code blocks are placed, and changed the way S0 and S1 are (aligned|adjusted)
in the less_than_4_ywords_before_S1_crosses_page_boundary logic.
After that changes, there is no significant performance change.
The ft_strcmp implementation slightly outperform the std when both strings can be aligned
and is equivalent to the std when they cannot.